Build process manager initializer for linux and windows
