+++ internal/node/processmanager/firecracker_procman.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package processmanager
 
 import (
